ENH: upgrade django-registration
Changes:

1. move built-in registration.backends.default.urls into person.urls (from
scipy_central.urls to scipy_central.person.urls)

2. django-registration latest version NOT compatible with previous
versions. Its majorly written from scratch using class-based `views`
(registration.backends.default.views.RegistrationView)

3. SciPyRegistrationBackend() is inherited from class-based view [2] and
moved into views.py

required packages:
django-registration==1.0
