+++ requirements.txt
@@ -2,7 +2,7 @@ Django==1.5
 django-haystack==2.1.0
 Whoosh==2.5.1
 South
-django-registration ==0.8
+django-registration==1.0
 django-widget-tweaks
 Sphinx
 Pygments
