+++ scipy_central/person/views.py
@@ -1,11 +1,13 @@
 from django.contrib.auth.decorators import login_required
+from django.contrib.auth.models import User
+from django.contrib.sites.models import Site
+from django.contrib.sites.models import RequestSite
 from django.shortcuts import render_to_response, redirect
 from django.template import RequestContext
 from django.core.exceptions import ObjectDoesNotExist
 from django.conf import settings
 from django.template.loader import render_to_string
 from django.db.utils import IntegrityError
-
 
 # Imports from other SciPy Central apps
 from scipy_central.pages.views import page_404_error
@@ -15,9 +17,16 @@ from scipy_central.utils import paginated_queryset, send_email
 from scipy_central.tagging.views import get_and_create_tags
 from scipy_central.pagehit.views import create_hit
 
+# 3rd party imports
+from registration.backends.default.views import RegistrationView
+from registration.models import RegistrationProfile
+from registration import signals
+
+# local package imports
 import models
 import forms
 
+# python
 import random
 import logging
 
@@ -219,3 +228,63 @@ def account_activation(user, **kwargs):
         rev.is_displayed = True
         rev.validation_hash = None
         rev.save()
+
+class SciPyRegistrationBackend(RegistrationView):
+
+    form_class = forms.SignUpForm
+
+    def register(self, request, **kwargs):
+        """
+        Given a username, email address and password, register a new
+        user account, which will initially be inactive.
+
+        Along with the new ``User`` object, a new
+        ``registration.models.RegistrationProfile`` will be created,
+        tied to that ``User``, containing the activation key which
+        will be used for this account.
+
+        An email will be sent to the supplied email address; this
+        email should contain an activation link. The email will be
+        rendered using two templates. See the documentation for
+        ``RegistrationProfile.send_activation_email()`` for
+        information about these templates and the contexts provided to
+        them.
+
+        After the ``User`` and ``RegistrationProfile`` are created and
+        the activation email is sent, the signal
+        ``registration.signals.user_registered`` will be sent, with
+        the new ``User`` as the keyword argument ``user`` and the
+        class of this backend as the sender.
+
+        """
+        username, email, password = kwargs['username'], kwargs['email'], kwargs['password1']
+        if Site._meta.installed:
+            site = Site.objects.get_current()
+        else:
+            site = RequestSite(request)
+
+        # We are creating a user with the same email address. We have already
+        # verified in ``forms.py`` that this isn't a mistake. Go ahead and pull
+        # the existing user from the DB and return that user instead.
+        if User.objects.filter(email__iexact=email):
+            new_user = User.objects.filter(email__iexact=email)[0]
+            new_user.username = username
+            new_user.set_password(password)
+            new_user.save()
+
+            # Resave their profile also (updates the slug)
+            new_user_profile = UserProfile.objects.get(user=new_user)
+            new_user_profile.save()
+
+            # Complete the activation email part
+            registration_profile = RegistrationProfile.objects.create_profile(new_user)
+            registration_profile.send_activation_email(site)
+        else:
+
+            new_user = RegistrationProfile.objects.create_inactive_user(\
+                                               username, email,password, site)
+
+        signals.user_registered.send(sender=self.__class__,
+                                     user=new_user,
+                                     request=request)
+        return new_user
