+++ scipy_central/urls.py
@@ -7,7 +7,7 @@ urlpatterns = patterns('scipy_central',
     # Major pages in the site: front page, about page, search, etc
     url(r'', include('scipy_central.pages.urls')),
 
-    # User authentication and profile viewing.
+    # User registration, authentication and profile viewing.
     url(r'^user/', include('scipy_central.person.urls'), ),
 
     # Submissions: new and existing, including previous revisions
@@ -15,11 +15,6 @@ urlpatterns = patterns('scipy_central',
 
     # reST comment converted to HTML
     url(r'rest/', include('scipy_central.rest_comments.urls')),
-
-    # Django-registration: new accounts, password resets, etc
-    # NOTE: the default backend is overriden ONLY for new account registration
-    # in scipy_central.person.urls
-    (r'^user/', include('registration.backends.default.urls')),
 
     # Tagging
     (r'^tagging/', include('scipy_central.tagging.urls')),
